requirements.txt
openai
python-dotenv
pydantic
pydantic-settings
asyncio
aiohttp